Move mdx-images to basic
