+++ packages/integrations/mdx/test/fixtures/basic/astro.config.mjs
@@ -4,5 +4,8 @@ import mdx from "@astrojs/mdx";
 
 export default defineConfig({
   site: 'https://mdx-is-neat.com/',
-  integrations: [react(), mdx()]
+  integrations: [react(), mdx()],
+  experimental: {
+		assets: true
+	}
 });
