+++ packages/integrations/mdx/test/fixtures/basic/src/assets/houston.webp
