+++ packages/integrations/mdx/test/fixtures/basic/src/pages/images/index.mdx
@@ -1,5 +1,5 @@
 Image using a relative path:
-![Houston](../assets/houston.webp)
+![Houston](../../assets/houston.webp)
 
 Image using an aliased path:
 ![Houston](~/assets/houston.webp)
