--- packages/integrations/mdx/test/fixtures/mdx-images/astro.config.ts
@@ -1,8 +1,0 @@
-import mdx from '@astrojs/mdx';
-
-export default {
-	integrations: [mdx()],
-	experimental: {
-		assets: true
-	}
-}
