--- packages/integrations/mdx/test/fixtures/mdx-images/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@test/mdx-images",
-  "private": true,
-  "dependencies": {
-    "@astrojs/mdx": "workspace:*",
-    "astro": "workspace:*",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  }
-}
